requirements.txt
lightning==2.4.0
pandas==2.2.3
pydub==0.25.1
mutagen==1.47.0
scikit-learn~=1.6.0rc1
pytorch_lightning~=2.4.0
torch~=2.8.0
pytorch-lightning~=2.4.0
torchmetrics~=1.8.1
numpy~=2.2.6
PyAudio~=0.2.14
PyAudioWPatch~=0.2.12.7
torchaudio~=2.8.0
matplotlib~=3.10.5
seaborn~=0.13.2
kagglehub~=0.3.12
requests~=2.32.4
tensorboard~=2.20.0
customtkinter~=5.2.2
onnx~=1.18.0
pillow~=11.0.0
onnxruntime~=1.22.1